converted types to architecture changeable

+++ arch/x86_64/h/linked_list.h
@@ -3,7 +3,7 @@
 #define _linked_list_h
 
 #include <stdint.h>
-
+#include <stddef.h>
 typedef struct _list_node_t
 {
     struct _list_node_t *prev;
@@ -12,7 +12,7 @@ typedef struct _list_node_t
 
 typedef struct
 {
-    uint64_t count;
+    size_t count;
     list_node_t list;
 }list_head_t;
 
@@ -24,5 +24,5 @@ list_node_t *linked_list_first(list_head_t *lh);
 list_node_t *linked_list_next(list_node_t *ln);
 list_node_t *linked_list_last(list_head_t *lh);
 list_node_t *linked_list_prev(list_node_t *ln);
-uint64_t linked_list_count(list_head_t *lh);
+size_t linked_list_count(list_head_t *lh);
 #endif
